+++ packages/custom-webpack/src/custom-webpack-builder.spec.ts
@@ -1,9 +1,11 @@
 import { Path } from '@angular-devkit/core';
 import { Configuration } from 'webpack';
+import { CustomizeRule } from 'webpack-merge';
 
 import { CustomWebpackBuilder, defaultWebpackConfigPath } from './custom-webpack-builder';
-import { MergeStrategies } from './custom-webpack-builder-config';
+import { MergeRules } from './custom-webpack-builder-config';
 import * as webpackConfigMerger from './webpack-config-merger';
+import { TargetOptions } from './type-definition';
 
 const baseWebpackConfig = {
   entry: './main.ts',
@@ -13,10 +15,10 @@ const buildOptions = {
   env: 'prod',
 };
 
-const targetOptions = {
+const targetOptions: TargetOptions = {
   project: 'application',
   configuration: 'production',
-  target: 'serve'
+  target: 'serve',
 };
 
 const customWebpackConfig = {
@@ -69,7 +71,7 @@ describe('CustomWebpackBuilder', () => {
       null,
       baseWebpackConfig,
       {},
-      {}
+      targetOptions
     );
 
     expect(mergedConfig).toEqual(baseWebpackConfig);
@@ -78,7 +80,13 @@ describe('CustomWebpackBuilder', () => {
   it('should load webpack.config.js if no path specified', async () => {
     const spy = jest.spyOn(webpackConfigMerger, 'mergeConfigs');
     createConfigFile(defaultWebpackConfigPath, customWebpackConfig);
-    await CustomWebpackBuilder.buildWebpackConfig(__dirname as Path, {}, baseWebpackConfig, {}, {});
+    await CustomWebpackBuilder.buildWebpackConfig(
+      __dirname as Path,
+      {},
+      baseWebpackConfig,
+      {},
+      targetOptions
+    );
 
     try {
       expect(spy).toHaveBeenCalledWith(
@@ -102,7 +110,7 @@ describe('CustomWebpackBuilder', () => {
       { path: fileName },
       baseWebpackConfig,
       {},
-      {}
+      targetOptions
     );
 
     try {
@@ -120,21 +128,21 @@ describe('CustomWebpackBuilder', () => {
   it('should pass on merge strategies', async () => {
     const spy = jest.spyOn(webpackConfigMerger, 'mergeConfigs');
     createConfigFile(defaultWebpackConfigPath, customWebpackConfig);
-    const mergeStrategies: MergeStrategies = { blah: 'prepend' };
+    const mergeRules: MergeRules = { blah: CustomizeRule.Prepend };
 
     await CustomWebpackBuilder.buildWebpackConfig(
       __dirname as Path,
-      { mergeStrategies },
+      { mergeRules },
       baseWebpackConfig,
       {},
-      {}
+      targetOptions
     );
 
     try {
       expect(spy).toHaveBeenCalledWith(
         baseWebpackConfig,
         customWebpackConfig,
-        mergeStrategies,
+        mergeRules,
         undefined
       );
     } finally {
@@ -151,7 +159,7 @@ describe('CustomWebpackBuilder', () => {
       { replaceDuplicatePlugins: true },
       baseWebpackConfig,
       {},
-      {}
+      targetOptions
     );
 
     try {
@@ -162,7 +170,7 @@ describe('CustomWebpackBuilder', () => {
   });
 
   it('should pass build options to the webpack config function', async () => {
-    const spy = jest.fn((config, options) => config);
+    const spy = jest.fn((config, options, targetOptions) => config);
     createConfigFile(defaultWebpackConfigPath, spy);
     await CustomWebpackBuilder.buildWebpackConfig(
       __dirname as Path,
@@ -182,7 +190,7 @@ describe('CustomWebpackBuilder', () => {
       {},
       baseWebpackConfig,
       {},
-      {}
+      targetOptions
     );
 
     expect(mergedConfig).toEqual(customWebpackFunctionObj);
@@ -206,7 +214,7 @@ describe('CustomWebpackBuilder', () => {
       {},
       baseWebpackConfig,
       {},
-      {}
+      targetOptions
     );
 
     expect(result).toBeInstanceOf(Promise);
@@ -235,7 +243,7 @@ describe('CustomWebpackBuilder', () => {
       {},
       baseWebpackConfig,
       {},
-      {}
+      targetOptions
     );
 
     expect(result).toBeInstanceOf(Promise);