+++ packages/custom-webpack/src/utils.ts
@@ -3,13 +3,24 @@
  * @param file: file name or file directory are allowed
  * @todo tsNodeRegistration: require ts-node if file extension is TypeScript
  */
-export function tsNodeRegister(file: string = '') {
+export function tsNodeRegister(file: string = '', tsConfig?: string) {
   if (file && file.endsWith('.ts')) {
     // Register TS compiler lazily
     require('ts-node').register({
+      project: tsConfig,
       compilerOptions: {
-        module: 'commonjs',
+        module: 'CommonJS',
+        types: [
+          'node', // NOTE: `node` is added because users scripts can also use pure node's packages as webpack or others
+        ],
       },
     });
+
+    // Register paths in tsConfig
+    const tsconfigPaths = require('tsconfig-paths');
+    const { absoluteBaseUrl: baseUrl, paths } = tsconfigPaths.loadConfig(tsConfig);
+    if (baseUrl && paths) {
+      tsconfigPaths.register({ baseUrl, paths });
+    }
   }
 }