+++ packages/jest/CHANGELOG.md
@@ -3,6 +3,118 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [12.1.0](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@12.1.0-beta.0...@angular-builders/jest@12.1.0) (2021-06-08)
+
+**Note:** Version bump only for package @angular-builders/jest
+
+## [12.1.0-beta.0](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@12.0.0...@angular-builders/jest@12.1.0-beta.0) (2021-06-07)
+
+### Features
+
+- **jest:** update to jest-preset-angular@9 ([#913](https://github.com/just-jeb/angular-builders/issues/913)) ([6bd90f8](https://github.com/just-jeb/angular-builders/commit/6bd90f898289466832f91cde9e1304955a7d43e3))
+
+## [12.0.0](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@12.0.0-beta.0...@angular-builders/jest@12.0.0) (2021-05-21)
+
+**Note:** Version bump only for package @angular-builders/jest
+
+## [12.0.0-beta.0](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@11.2.0...@angular-builders/jest@12.0.0-beta.0) (2021-05-14)
+
+### ⚠ BREAKING CHANGES
+
+- **deps:** update to angular 12
+
+- chore(deps): update to node 12.14 on CI
+
+- Update html-webpack-plugin in example
+
+### Miscellaneous Chores
+
+- **deps:** update to angular 12 ([#980](https://github.com/just-jeb/angular-builders/issues/980)) ([ad2772b](https://github.com/just-jeb/angular-builders/commit/ad2772b33d6b62806e791dd908696945a77e7a98)), closes [#972](https://github.com/just-jeb/angular-builders/issues/972)
+
+## [11.2.0](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@11.2.0-beta.1...@angular-builders/jest@11.2.0) (2021-04-20)
+
+**Note:** Version bump only for package @angular-builders/jest
+
+## [11.2.0-beta.1](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@11.2.0-beta.0...@angular-builders/jest@11.2.0-beta.1) (2021-04-20)
+
+### Bug Fixes
+
+- **jest:** correct maxWorkers type ([#970](https://github.com/just-jeb/angular-builders/issues/970)) ([eddb4db](https://github.com/just-jeb/angular-builders/commit/eddb4db9d1de883eb937549ce27866b59eb120cb))
+
+## [11.2.0-beta.0](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@11.1.1...@angular-builders/jest@11.2.0-beta.0) (2021-04-13)
+
+### Features
+
+- **jest:** support configurable global mocks ([#965](https://github.com/just-jeb/angular-builders/issues/965)) ([ba6af13](https://github.com/just-jeb/angular-builders/commit/ba6af1308e6782bba873bce4236abf9251124345))
+
+### [11.1.1](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@11.1.1-beta.0...@angular-builders/jest@11.1.1) (2021-03-09)
+
+**Note:** Version bump only for package @angular-builders/jest
+
+### [11.1.1-beta.0](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@11.1.0...@angular-builders/jest@11.1.1-beta.0) (2021-03-08)
+
+### Bug Fixes
+
+- **jest:** sync latest jest-preset-angular ([805eb8d](https://github.com/just-jeb/angular-builders/commit/805eb8d0d0dbe37582a32621e0239d0ee107e6b9))
+
+## [11.1.0](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@11.1.0-beta.0...@angular-builders/jest@11.1.0) (2021-02-26)
+
+**Note:** Version bump only for package @angular-builders/jest
+
+## 11.1.0-beta.0 (2021-02-03)
+
+### Features
+
+- **custom-webpack:** use application tsconfig file for indexTransform and customWebpackConfig ([#879](https://github.com/just-jeb/angular-builders/issues/879)) ([c85cd10](https://github.com/just-jeb/angular-builders/commit/c85cd103af6047fefbea273ace1d9446829f3651))
+
+### 11.0.1-beta.4 (2021-02-03)
+
+**Note:** Version bump only for package @angular-builders/jest
+
+### 11.0.1-beta.3 (2021-02-03)
+
+**Note:** Version bump only for package @angular-builders/jest
+
+### 11.0.1-beta.2 (2021-02-03)
+
+**Note:** Version bump only for package @angular-builders/jest
+
+### 11.0.1-beta.1 (2021-01-14)
+
+**Note:** Version bump only for package @angular-builders/jest
+
+### 11.0.1-beta.0 (2020-12-27)
+
+**Note:** Version bump only for package @angular-builders/jest
+
+## [11.0.0](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@11.0.0-beta.6...@angular-builders/jest@11.0.0) (2020-12-27)
+
+**Note:** Version bump only for package @angular-builders/jest
+
+## 11.0.0-beta.6 (2020-12-27)
+
+**Note:** Version bump only for package @angular-builders/jest
+
+## 11.0.0-beta.5 (2020-12-22)
+
+### Bug Fixes
+
+- **custom-webpack:** allow merging loader name with loader object ([#912](https://github.com/just-jeb/angular-builders/issues/912)) ([a89a35d](https://github.com/just-jeb/angular-builders/commit/a89a35d30276a3d284d304ee710442ee03da351b))
+
+## 11.0.0-beta.4 (2020-12-16)
+
+### Bug Fixes
+
+- **custom-webpack:** fix loaders merge scenario ([#909](https://github.com/just-jeb/angular-builders/issues/909)) ([0f0ad3d](https://github.com/just-jeb/angular-builders/commit/0f0ad3d174ea589538c4474b0509ad3b5f925bf1))
+
+## 11.0.0-beta.3 (2020-12-15)
+
+**Note:** Version bump only for package @angular-builders/jest
+
+## 11.0.0-beta.2 (2020-12-14)
+
+**Note:** Version bump only for package @angular-builders/jest
+
 ## [11.0.0-beta.1](https://github.com/just-jeb/angular-builders/compare/@angular-builders/jest@10.0.2-beta.1...@angular-builders/jest@11.0.0-beta.1) (2020-11-17)
 
 ### ⚠ BREAKING CHANGES